+++ app/database/tables/portfolio/get_sector_breakdown.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from app.database.helpers.fetch_all import fetch_all
 
 def get_sector_breakdown(portfolio_id):
@@ -35,8 +34,8 @@ def get_sector_breakdown(portfolio_id):
     args = (portfolio_id,)
     sector_data = fetch_all(query=query, args=args, dictionary=True)
 
-    # Convert values to Decimal to avoid scientific notation
-    breakdown = {f"{item['sector'].lower()}_value": Decimal(item['total_value'] or 0) for item in sector_data}
+    # Convert values to float to avoid TypeError in templates
+    breakdown = {f"{item['sector'].lower()}_value": float(item['total_value'] or 0) for item in sector_data}
 
     # Calculate total portfolio value
     total_value = sum(breakdown.values())
